Added user authorization

+++ backend/routes/posts.routes.js
@@ -64,7 +64,8 @@ router.post(
     const post = new Post({
       title: req.body.title,
       content: req.body.content,
-      imagePath: url + "/images/" + req.file.filename
+      imagePath: url + "/images/" + req.file.filename,
+      creator: req.userData.userId
     });
     post.save().then(createdPost => {
       res.status(201).json({
@@ -93,13 +94,22 @@ router.put(
       _id: req.body.id,
       title: req.body.title,
       content: req.body.content,
-      imagePath: imagePath
+      imagePath: imagePath,
+      creator: req.userData.userId
     });
-    console.log(post);
-    Post.updateOne({ _id: req.params.id }, post).then(result => {
-      res.status(200).json({
-        message: "Post updated successfully!"
-      });
+    Post.updateOne(
+      { _id: req.params.id, creator: req.userData.userId },
+      post
+    ).then(result => {
+      if (result.nModified > 0) {
+        res.status(200).json({
+          message: "Post updated successfully!"
+        });
+      } else {
+        res.status(401).json({
+          message: "Post updation failed, user not authorized!"
+        });
+      }
     });
   }
 );
@@ -117,11 +127,17 @@ router.get("/:id", (req, res, next) => {
 
 // Delete post/api/posts
 router.delete("/:id", checkAuth, (req, res, next) => {
-  Post.deleteOne({ _id: req.params.id })
+  Post.deleteOne({ _id: req.params.id, creator: req.userData.userId })
     .then(result => {
-      res.status(200).json({
-        message: "Post deleted!"
-      });
+      if (result.n > 0) {
+        res.status(200).json({
+          message: "Post deleted successfully!"
+        });
+      } else {
+        res.status(401).json({
+          message: "Post deletion failed, user not authorized!"
+        });
+      }
     })
     .catch(error => console.log(error));
 });
